+++ src/module/applications/sheets/_types.d.ts
@@ -70,3 +70,10 @@ export interface ActorSheetEquipmentContext {
   equipment: ActorSheetItemContext[];
   showAdd: boolean;
 }
+
+export interface ActorSheetComplicationsContext {
+  label: string;
+  complication: ActorSheetItemContext[];
+  showAdd: boolean;
+  showHeader: boolean;
+}