+++ CHANGELOG.md
@@ -25,10 +25,12 @@
 - Advancements (#51)
 - Added a button to ability use messages to apply effects from the ability. (#214)
 - Added a new Active Effect subtype for ability bonuses. (#713)
+- Added Epic Resource display. (#537)
 - Added a Wealth field to Careers. (#535)
 - Adding or removing a career will adjust the amount of renown and wealth your character has.
 - Added a Renown and Wealth field to characters. (#279)
 - Added age, height, and weight info to the Biography tab for characters.
+- Added data path for bonuses to max stamina per echelon. (#259)
 - Expanded end of encounter prompt to reset heroic resources & temporary stamina. (#612)
 - Added name and image inputs to pseudodocument creation, e.g. Power Roll Effects. (#718)
 - Added a "Next Respite" option for ending Active Effects. (#694)